requirements.txt
click==8.3.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
blinker==1.9.0
Flask==3.1.2
Werkzeug==3.1.4
python-dotenv
requests




